# educa/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Educa LRS REST API Backend",
        "description": "Educa LRS API Backend",
        "contact": {
            "email": "support@digitallearning.gmbh"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/statements/filter": {
            "get": {
                "tags": [
                    "Statements"
                ],
                "summary": "Filter statements based on criteria",
                "operationId": "2ca638d5e3976166fb8f896fb956473d",
                "parameters": [
                    {
                        "name": "actor_name",
                        "in": "query",
                        "description": "Filter statements by actor name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "John Doe"
                        }
                    },
                    {
                        "name": "verb_name",
                        "in": "query",
                        "description": "Filter statements by verb name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "completed"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Filtered list of statements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "actor": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "mbox": {
                                                        "type": "string",
                                                        "example": "mailto:john.doe@example.com"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "verb": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "http://adlnet.gov/expapi/verbs/completed"
                                                    },
                                                    "display": {
                                                        "properties": {
                                                            "en-US": {
                                                                "type": "string",
                                                                "example": "completed"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "object": {
                                                "properties": {
                                                    "objectType": {
                                                        "type": "string",
                                                        "example": "Activity"
                                                    },
                                                    "id": {
                                                        "type": "string",
                                                        "example": "http://example.com/activities/module-1"
                                                    },
                                                    "definition": {
                                                        "properties": {
                                                            "name": {
                                                                "properties": {
                                                                    "en-US": {
                                                                        "type": "string",
                                                                        "example": "Module 1"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            },
                                                            "description": {
                                                                "properties": {
                                                                    "en-US": {
                                                                        "type": "string",
                                                                        "example": "A module description"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}